feat : 쿠폰 할인 금액 타입 변경

+++ src/main/java/com/bookservice/coupon/dto/request/CouponRegisterRequest.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 public class CouponRegisterRequest {
 	private String name;
 	private Integer amount;
-	private Integer discountValue;
+	private Long discountValue;
 
 	public Coupon toCoupon() {
 		return Coupon.builder()
